requirements.txt
streamlit
python-docx
PyPDF2
nltk
numpy
scikit-learn